Create README.md
This project is a Pomodoro Timer, a time management tool that breaks down work into intervals separated by breaks. It was built using JavaScript, with the goal of accurately counting down the minutes and seconds and updating the timer display on the page. The setInterval() function and padStart() method were used to overcome technical challenges and enhance the user experience. The timer includes start and stop buttons and a settings button that allows the user to change the length of the timer. This was a challenging but rewarding project that required careful consideration of logic and user experience. Feedback and suggestions are welcome.